New Transposition Cipher added
Added a new Transposition Cipher. You can encrypt and decrypt some data
either by reading from a file or direct through console. After the
encrypt/decrypt action is performed you can choose to write the results
to a file or just display to console.

Added a new method in the caesar_cipher module to crack the cipher if
the key is unknown